Refactor: Added index file to ease imports of multiple components in a single import statement
