+++ src/components/index.js
@@ -1,0 +1,25 @@
+import Back from "./Back/Back"
+import DatePickerItem from "./DatePicker/DatePicker"
+import FavoriteMovie from "./FavoriteMovie/FavoriteMovie"
+import Loading from "./Loading/Loading"
+import Movie from "./Movie/Movie"
+import Navbar from "./Navbar/Navbar"
+import NoMovies from "./NoMovies/NoMovies"
+import NoSearchItems from "./NoSearchItems/NoSearchItems"
+import NotFoundMovie from "./NotFoundMovie/NotFoundMovie"
+import ProtectedRoute from './ProtectedRoute/ProtectedRoute'
+import SortControls from "./SortControls/SortControls"
+
+export {
+    Back,
+    DatePickerItem,
+    FavoriteMovie,
+    Loading,
+    Movie,
+    Navbar,
+    NoMovies,
+    NoSearchItems,
+    NotFoundMovie,
+    ProtectedRoute,
+    SortControls
+}
