+++ src/pages/AddMovie/AddMovie.jsx
@@ -1,13 +1,11 @@
 import React, { useState } from 'react'
 import './AddMovie.scss'
 import { createMovie } from '../../api'
-import Back from '../../components/Back/Back'
 import { useNavigate } from 'react-router-dom'
-import DatePickerItem from '../../components/DatePicker/DatePicker'
 import { useDispatch, useSelector } from 'react-redux'
 import { addMovie } from '../../redux/slices/moviesSlice'
 import Rating from '@mui/material/Rating'
-import FavoriteMovie from '../../components/FavoriteMovie/FavoriteMovie'
+import { DatePickerItem, FavoriteMovie, Back } from '../../components'
 import placeholderImg from '../../assets/placeholder2.jpeg'
 import ClearIcon from '@mui/icons-material/Clear'
