+++ src/pages/index.js
@@ -1,7 +1,23 @@
 import Login from "./Login/Login"
 import Register from "./Register/Register"
+import AddMovie from './AddMovie/AddMovie'
+import DeleteMovie from "./DeleteMovie/DeleteMovie"
+import EditMovie from './EditMovie/EditMovie'
+import Layout from './Layout/Layout'
+import MovieDetail from './MovieDetail/MovieDetail'
+import Movies from './Movies/Movies'
+import NotFound from './NotFound/NotFound'
+import Profile from './Profile/Profile'
 
 export {
     Login,
-    Register
+    Register,
+    AddMovie,
+    DeleteMovie,
+    EditMovie,
+    Layout,
+    MovieDetail,
+    Movies,
+    NotFound,
+    Profile
 }
